Refactor GENET device code for improved type safety and consistency
- Replaced custom strcmp implementation with _Stricmp from strutil.h for string comparison in phy_interface.c.
- Changed loop index types from int to u8 in unit.c for better type safety.
- Updated return types of several functions in unit.c and unit_commands.c from int to u32 to reflect correct usage.
- Modified GENET_SupportedCommands array type from UWORD to u16 in unit_commands.c for consistency.
- Adjusted function signatures and variable types in unit_commands_mcast.c to use u64 and u32 for better clarity and type safety.
- Enhanced CopyPacket function in unit_io.c to use u8 and u32 types, ensuring proper handling of packet data.
- Updated runtime configuration types in runtime_config.h and runtime_config.c for better type consistency and safety.
- Added additional compiler warnings in CMakeLists.txt to catch potential issues during compilation.

+++ genet.device/include/genet/bcmgenet-irq.h
@@ -2,12 +2,13 @@
 #ifndef BCMGENET_IRQ_H
 #define BCMGENET_IRQ_H
 
+#include <types.h>
 #include <genet/bcmgenet-regs.h>
 #include <device.h>
 
 /* Interrupt enable/disable */
-void bcmgenet_irq0_enable(struct GenetUnit *unit, ULONG irq_mask);
-void bcmgenet_irq0_disable(struct GenetUnit *unit, ULONG irq_mask);
+void bcmgenet_irq0_enable(struct GenetUnit *unit, u32 irq_mask);
+void bcmgenet_irq0_disable(struct GenetUnit *unit, u32 irq_mask);
 void bcmgenet_intr_disable(struct GenetUnit *unit);
 
 /* Interrupt handler */
